Adds static scheduler and basic user programs printing information via UART.

+++ inc/types.h
@@ -1,8 +1,7 @@
 // See LICENSE file for copyright and license details.
-
 #pragma once
 
-typedef unsigned long uintptr_t;
+typedef unsigned long long uintptr_t;
 typedef unsigned char uint8_t;
 typedef unsigned long long uint64_t;
 typedef unsigned long size_t;
